requirements.txt
Flask==3.0.3
PyMuPDF==1.24.10
openai==1.3.0
httpx==0.25.2
python-dotenv==1.0.1
